third_party/WebKit/LayoutTests/platform/win-xp/svg/animations/animate-elem-06-t-drt-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests


A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.getTransformToElement(path).e is -15
PASS rootSVGElement.getTransformToElement(path).f is -43
FAIL rootSVGElement.getTransformToElement(path).e should be close to -38.5. Was -38.3656120300293.
PASS rootSVGElement.getTransformToElement(path).f is -30
PASS rootSVGElement.getTransformToElement(path).e is -65
PASS rootSVGElement.getTransformToElement(path).f is -33
PASS rootSVGElement.getTransformToElement(path).e is -65
PASS rootSVGElement.getTransformToElement(path).f is -33
PASS successfullyParsed is true

TEST COMPLETE
